refactor(ui): update structure for select and multiselect components
